add result object

+++ src/features/comments/comments-service.ts
@@ -1,15 +1,45 @@
 import {commentsRepository} from "../comments";
-import {TInputComment} from "../../types";
+import {TInputComment, TResultServiceObj} from "../../types";
 import {TCommentDB} from "../../db";
+import {createServiceResultObj} from "../../utils";
 
 export const commentsService = {
-    async findCommentById(id: string): Promise<TCommentDB | null> {
-        return await commentsRepository.findCommentById(id);
+    async findCommentById(id: string): Promise<TResultServiceObj<TCommentDB>> {
+        const comment = await commentsRepository.findCommentById(id);
+        if (comment) {
+            return createServiceResultObj<TCommentDB>("SUCCESS", "OK", comment);
+        }
+
+        return createServiceResultObj("SUCCESS", "NOT_FOUND");
     },
-    async updateCommentById(id: string, data: TInputComment): Promise<boolean> {
-        return await commentsRepository.updateCommentById(id, data);
+    async updateCommentById(userId: string, id: string, data: TInputComment): Promise<TResultServiceObj> {
+        const {result, status, data: findCommentByIdData} = await this.findCommentById(id);
+
+        if (result === "SUCCESS") {
+            if (findCommentByIdData!.commentatorInfo.userId === userId) {
+                await commentsRepository.updateCommentById(id, data);
+
+                return createServiceResultObj("SUCCESS", "NO_CONTENT");
+            }
+
+            return createServiceResultObj("REJECT", "FORBIDDEN")
+        }
+
+        return createServiceResultObj(result, status);
     },
-    async deleteCommentById(id: string): Promise<boolean> {
-        return await commentsRepository.deleteCommentById(id);
+    async deleteCommentById(userId: string, id: string): Promise<TResultServiceObj> {
+        const {result, status, data} = await this.findCommentById(id);
+
+        if (result === "SUCCESS") {
+            if (data!.commentatorInfo.userId === userId) {
+                await commentsRepository.deleteCommentById(id);
+
+                return createServiceResultObj("SUCCESS", "NO_CONTENT");
+            }
+
+            return createServiceResultObj("REJECT", "FORBIDDEN")
+        }
+
+        return createServiceResultObj(result, status);
     }
 }
